fix: raise proper ValueError instead of SIGSEGV

+++ pillow_heif/_pillow_heif.c
@@ -155,7 +155,7 @@ static PyObject* _CtxWriteImage_add_plane(CtxWriteImageObject* self, PyObject* a
         real_stride = real_stride * 2;
     if (stride_in == 0)
         stride_in = real_stride;
-    if (stride_in * height > buffer.len) {
+    if ((Py_ssize_t)stride_in * height > buffer.len) {
         PyBuffer_Release(&buffer);
         PyErr_SetString(PyExc_ValueError, "image plane does not contain enough data");
         return NULL;
@@ -341,7 +341,7 @@ static PyObject* _CtxWriteImage_add_plane_la(CtxWriteImageObject* self, PyObject
         real_stride = real_stride * 2;
     if (stride_in == 0)
         stride_in = real_stride;
-    if (stride_in * height > buffer.len) {
+    if ((Py_ssize_t)stride_in * height > buffer.len) {
         PyBuffer_Release(&buffer);
         PyErr_SetString(PyExc_ValueError, "image plane does not contain enough data");
         return NULL;
@@ -446,7 +446,7 @@ static PyObject* _CtxWriteImage_add_plane_l(CtxWriteImageObject* self, PyObject*
         real_stride = real_stride * 2;
     if (stride_in == 0)
         stride_in = real_stride;
-    if (stride_in * height > buffer.len) {
+    if ((Py_ssize_t)stride_in * height > buffer.len) {
         PyBuffer_Release(&buffer);
         PyErr_SetString(PyExc_ValueError, "image plane does not contain enough data");
         return NULL;
